+++ packages/asc-ui/src/components/FormTitle/FormTitleStyle.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import { themeColor, themeSpacing } from '../../utils/themeUtils'
+
+export default styled.h1`
+  max-width: 620px;
+  margin: ${themeSpacing(6, 0, 8)};
+  padding-bottom: 6px;
+  border-bottom: 1px solid ${themeColor('tint', 'level5')};
+  color: ${themeColor('tint', 'level5')};
+  font-size: 14px;
+  font-weight: 700;
+  line-height: 18px;
+  -webkit-font-smoothing: auto;
+`